#3 Enter boot loader mode on double click on the reset button
After reset a magic number is stored in a backup register that is not cleared on reset. After 500ms the register is set to 0 and the firmware is booted.
On the other hand, if reset is pressed before the 500ms, the magic number is still in the register. When the system is booted and the magic number is in the register, we enter boot loader mode instead.

Also cleaned up imports HAL files.